requirements.txt
pandas~=2.3
scipy~=1.16.3
numpy~=2.4.0
scikit-learn~=1.8.0
plotly~=6.5.0
requests~=2.32.5
PyYAML~=6.0.3
python-dotenv~=1.2.1